refactor(bayn): split candidate development evidence
